bella_core/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psutil
